Use janus-idp image and installation procedure
The charts is updated to use the recommended installation method of
janus-idp. By that method, dynamic plugins (orchestrator for now,
notification in the future) are installed via a configuration only
without a need to build a custom image to include them.

The plugins for now will be installed from npmjs registry under
caponetto account. In the future this will be updated to be install from
janus-idp official plugins.

The image for janus-idp relies on a specific build (pr-814), since
'latest' image is outdated and 'next' tag was broken at the time of
creating this PR.

Signed-off-by: Moti Asayag <masayag@redhat.com>