orchestrator-k8s: load locations from a local fs
Loading backstage locations from URL is limiting when all you need is to
be up and running with notifications plugins working and you don't want
the extra steps to configure a github token just to be able to pull those in
Instead the chart would install those locations files in a config map from
the `/locations` directory.
The price to pay is that those location files need to manually downloaded
and placed in the location dir from time to time.

Note: the location file in the values.yaml must match the name in the config map

To use the mount in backstage there's a new `backstage-location`
config map and an `backstage.extraVolumes` and `backstage.extraVolumeMounts`
entries in values.yaml

Signed-off-by: Roy Golan <rgolan@redhat.com>