Add support for configuring the notifications email server
The notifications email plugin enabled the user to fan-out emails if the
recipients has an email address associated with the user definition.
To enable this feature, there is a need to configure the notifications
email plugin.

The values.yaml exposes the values for some of the plugin attributed,
and the sensitive information (hostname, username and password) are
expected to be provided when executing the setup.sh script.

Signed-off-by: Moti Asayag <masayag@redhat.com>

rh-pre-commit.version: 2.3.0
rh-pre-commit.check-secrets: ENABLED