Add a variable for specify the software template branch
It is common use case to refer a specific branch for importing the
software templates from when developing features for the chart.
Instead of manipulating the chart locally, a variable is expose to ease
the experience when a different branch then the 'main' is needed.

Signed-off-by: Moti Asayag <masayag@redhat.com>

rh-pre-commit.version: 2.3.1
rh-pre-commit.check-secrets: ENABLED